drbd: Re-add primary nodes mask to struct p_twopc_request
In two-phase commits, the primary node mask is computed on the
coordinator node in phase one.  In phase two, if any primary nodes are
in the cluster, all nodes which are not connected to all of those
primary nodes must outdate themselves.  To allow that in all cases,
we must send the primary node mask around in P_TWOPC_COMMIT packets.