drbd: Move max_buffers to net_conf, mark transport_name invariant
max_buffers used to limit the io-depth on a device as well
as limit the number of buffer pages, that may be acllocated
on the receiving node.

The first use was eliminated, therefore move it to the
net_conf

transport_name should have been marked as INVARIANT in
the first place. It can not be changed while the connection
exists.