introduce P_CONFIRM_STABLE
which is supposed to resolve a potential distributed deadlock
if too many requests "scattered" over too many distinct LBA offsets
happen to be submitted within the same "epoch".

Where too many means: more than the activity log on the peer can hold,
which can be the case for differently configured al-extens on the peers,
or if the Primary is diskless (and thus has no activity log itself).