WinDRBD: New ioctl's and user device object
In WinDRBD 0.10.7, we've added new commands to the windrbd
utility. This patch adds the ioctl constants for the kernel
interface.

In addition there is now a user device object (accessible
by everyone) which allows drbdadm status and drbdsetup events2
to be executed by non-Administrators (just like in the Linux
DRBD variant).