fixup: remove backing device field from peer_device events
DRBD does not actually know about the backing devices of a peer.
As explained in the corresponding drbd commit:

It seems pretty logical that
peer_device->device->ldev->disk_conf->backing_dev would contain the
backing device of the disk on the peer. However, this is actually wrong:
that peer_device->device just points to the local device again.