Introduce res_fail_io status and on_susp_primary_outdated setting
res_fail_io reports to status and events2 when a node is failing IO
requests because it forcefully demoted a node to secondary.

on_susp_preimary_outdated is a new setting to fine tune the behaviour
when a primary that has IO suspended due to lack of quorum, suddenly
learns that an other primary exists, and that other primary has
started a more recent data generation.