WinDRBD: New ioctl code for allowing unloading the driver
Before that the driver automatically transistioned into a
state where it can be unloaded once usage count (bus devices
plus drbd devices) reaches 0. This prevented a followup
creation of the bus device to fail and later drbdadm primary
to stall (even tough there was a bus device, but the driver
didn't know). Now setting AddDevice to NULL is an explicit
step (unlock-driver) it should be called by the installer
just before the sc stop windrbd command.