drbd: introduce pointer and kref from path to connection
A path can not exist without a connection it is for. The lb-tcp and
the rdma transport implementation already had a pointer from their
path objects to the transport.

Moving that pointer to the generic path allows introducing to get a
reference on the transport/connection.

The 'verify-target-becomes-inconsistent' test on the lb-tcp transport
highlights that this change is necessary.