drbd: listeners need to hold a reference on the transport module
A listener object can live longer than a path or connection
objects. I.e. when it has a work item scheduled to carry out an
accept() call. We must ensure the module is not unloaded before that
work item is terminated. We can do that by holding references onto the
module from each listener.