transport: add return value to net_conf_change
In order to support more complex scenarios where a change to the net_conf
requires additional, fallible work by the transport, add a return value
to net_conf_change.

If negative, an error occured, and the net_conf was not applied.

As a second step, the caller should also call net_conf_change the
first time the transport gets created, right after it was initialized.