WinDRBD: Suspend/Resume I/O for a minor on the WinDRBD level (outside DRBD)
This adds two new ioctl() codes to the WinDRBD kernel interface.

When an application busy writes a block it may happen that
syncing never finished. In that case, set this suspend-io
flag, wait for sync to finish and then clear the suspend-io
flag again. Note that this is different from DRBD's suspend-io
command.