WinDRBD: New ioctl IOCTL_WINDRBD_ROOT_BUS_DEVICE_IS_WORKING
Sometimes after an installation of WinDRBD the Virtual
Bus Device does not exist. Since this device is crucial
for proper WinDRBD operation (else drbdadm primary will
not work) the installer asks the kernel module via this
ioctl if the bus device is present and working correctly

The ioctl returns an non-zero value if the bus device is
working correctly.